Rename domain:primary:set to domain:primary:add per consensus of product and technical leadership.
